Consolidates all code for the diagnosis example into a single file
